% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_sqmi.R
\name{area_sqmi}
\alias{area_sqmi}
\title{utility to get areas of places (points with radius at each, or polygons, or census units by FIPS)}
\usage{
area_sqmi(df = NULL, radius.miles = NULL, shp = NULL, fips = NULL)
}
\arguments{
\item{df}{optional data.frame, one place per row -
This function tries to infer sitetype by seeing if df is 1) a
spatial data.frame of class "sf", or 2) has a column that can be interpreted as
an alias for fips, or as last resort, 3) a column that is an alias for radius.miles}

\item{radius.miles}{optional vector of distances from points defining circular buffers}

\item{shp}{optional spatial data.frame sf class object like from \code{\link[=shapefile_from_any]{shapefile_from_any()}}}

\item{fips}{optional vector of character Census FIPS codes, with leading zeroes,
2 digits for State, 5 for county, etc. If you already have the boundaries
then provide that as shp instead of this parameter (much faster that way).}
}
\value{
vector of numbers same length as length(radius.miles) or length(fips) or NROW(shp)
}
\description{
utility to get areas of places (points with radius at each, or polygons, or census units by FIPS)
}
\details{
Only one of the parameters can be specified at a time, and others must be NULL.
If you provide a data.frame it tries to infer which info is in the table --
radius.miles, shp, or fips.

Note this is slow for fips since it has to download boundaries. If you already have
the shapefile of boundaries, provide that as the shp parameter instead of using fips.

Note: if you provide a single number for radius (a vector of length 1),
this returns a single value for area. If you provide a vector of radius values,
even if they are all the same number, this returns a vector as long as the input
radius.miles or as long as NROW(df).
}
\keyword{internal}
