% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROXIMITY_FUNCTIONS.R
\name{calc_counties_from_bg}
\alias{calc_counties_from_bg}
\title{DRAFT - Calculate (aggregate) county scores from block group scores}
\usage{
calc_counties_from_bg(
  childDT,
  score_colname,
  wt_colname = "pop",
  bgfips_colname = "bgfips",
  calc_method = c("wtdmean", "sum")[1]
)
}
\arguments{
\item{childDT}{data.table (or data.frame)}

\item{score_colname}{name of a column in childDT}

\item{wt_colname}{name of a column in childDT, used as weights for weighted mean of scores in each county}

\item{bgfips_colname}{name of a column in childDT, must be unique rows,
and first 5 characters must be the county FIPS code (and must include any leading zeroes)}
}
\value{
data.table of 1 row per county (each county that is in the childDT provided),
just columns "countyfips", "Countyname", score_colname, wt_colname
}
\description{
Redo as more generic and TO HANDLE >1 INDICATOR AT A TIME ! See other functions in PROXIMITY_FUNCTIONS.R !
}
\details{
This ignores any rows with NA in the score_colname, but
if you want an NA weight (in wt_colname) to count as a weight of 0, you have to convert them to zeroes first,
or this function will return NA any time there is any NA value at all in the wt_colname
}
\keyword{internal}
