% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROXIMITY_FUNCTIONS.R
\name{calctype}
\alias{calctype}
\title{utility - what type of formula is used to aggregate this variable?}
\usage{
calctype(varnames)
}
\arguments{
\item{varnames}{vector like names_d}
}
\value{
vector same length as varnames, like c("sum of counts", "wtdmean")
}
\description{
utility - what type of formula is used to aggregate this variable?
}
\examples{
 calctype("pop")
 calctype(names_d)
 
 x = names_these
 cbind(indicator = x, calctype = calctype(x), calcweight = calcweight(x))
 
 x = names(testoutput_ejamit_10pts_1miles$results_overall)
 cbind(indicator = x, calctype = calctype(x), calcweight = calcweight(x))
 
}
\keyword{internal}
