% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_view3_table.R
\name{create_view3_table}
\alias{create_view3_table}
\title{create_view3_table - function that creates site by site table in app}
\usage{
create_view3_table(data_processed, data_summarized, testing)
}
\arguments{
\item{data_processed}{processed data from app_server}

\item{data_summarized}{summarized data from app_server}

\item{testing}{variable to indicate if analysis is being tested}
}
\description{
Builds site by site table after an analysis in EJAM app. Pulls in uploaded and analyzed data to create table
}
