% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataload_dynamic.R
\name{dataload_dynamic}
\alias{dataload_dynamic}
\title{Utility to download / load datasets}
\usage{
dataload_dynamic(
  varnames = .arrow_ds_names[1:3],
  envir = globalenv(),
  folder_local_source = NULL,
  silent = FALSE,
  return_data_table = TRUE,
  onAttach = FALSE
)
}
\arguments{
\item{varnames}{character vector of names of R objects to get from board,
or set this to "all" to load all of them}

\item{envir}{if needed to specify environment other than default, e.g., globalenv() or parent.frame()}

\item{folder_local_source}{path of local folder to
look in for locally saved copies in case pins board is not reachable by user.}

\item{silent}{set to TRUE to suppress cat() msgs to console}

\item{return_data_table}{whether the \code{\link[=read_ipc_file]{read_ipc_file()}} should return a data.table (T, the default), or arrow (F). Passed to \code{\link[=dataload_from_local]{dataload_from_local()}}}

\item{onAttach}{Indicates whether the function is being called from onAttach. IF so, it will download all arrow files if user cannot connect to PINS board}
}
\value{
returns vector of names of objects now in memory in specified envir, either because
\enumerate{
\item already in memory or
\item loaded from local disk or
\item successfully downloaded.
}
}
\description{
Utility to download / load datasets
}
\details{
First checks memory, then installed package's data folder
When package is first loaded, all arrow files are downloaded from USEPA/ejamdata
into package's data directory
}
