% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barplot_ratios.R
\name{ejam2barplot}
\alias{ejam2barplot}
\alias{plot_barplot_ratios_ez}
\alias{plot_barplot_ratios}
\title{Barplot of ratios of residential population (or other) scores to averages - simpler syntax}
\usage{
ejam2barplot(
  ejamitout,
  varnames = c(names_d_ratio_to_avg, names_d_subgroups_ratio_to_avg),
  sitenumber = NULL,
  main = "Residential Populations at the Analyzed Locations Compared to US Overall",
  ...
)

plot_barplot_ratios_ez(
  out,
  varnames = c(names_d_ratio_to_avg, names_d_subgroups_ratio_to_avg),
  main = "Residential Populations at the Analyzed Locations Compared to US Overall",
  single_location = FALSE,
  row_index = NULL,
  ...
)

plot_barplot_ratios(
  ratio.to.us.d.overall,
  shortlabels = NULL,
  mycolorsavailable = c("gray", "yellow", "orange", "red"),
  main = "Residential Populations at the Analyzed Locations Compared to US Overall"
)
}
\arguments{
\item{ejamitout}{like from \code{\link[=ejamit]{ejamit()}}}

\item{varnames}{vector of indicator names that are ratios to avg, like
c(names_d_ratio_to_avg , names_d_subgroups_ratio_to_avg)
but could be c(names_d_ratio_to_state_avg , names_d_subgroups_ratio_to_state_avg)}

\item{sitenumber}{default is all sites from ejamitout$results_overall, and
if an integer, it is the site number to show from ejamitout$results_bysite}

\item{main}{title for plot, like "Stats at the Analyzed Locations Compared to US Overall"}

\item{...}{
  Arguments passed on to \code{\link[=ejam2barplot]{ejam2barplot}}
  \describe{
    \item{\code{}}{}
  }}

\item{out}{the list of tables that is the output of ejamit() or a related function}

\item{single_location}{set to TRUE if using row_index to view one site,
set to FALSE to view overall results from out$results_overall}

\item{row_index}{the number of the row to use from out$results_bysite}

\item{ratio.to.us.d.overall}{named list of a few ratios to plot, but see \code{\link[=ejam2barplot]{ejam2barplot()}}
for an easier way to specify which indicator to show.}

\item{shortlabels}{names to use for plot - should be same length as named list ratio.to.us.d.overall}

\item{mycolorsavailable}{leave as default}
}
\value{
ggplot

ggplot should be returned
}
\description{
Barplot of ratios of residential population (or other) scores to averages - simpler syntax

Same as ejam2barplot() but ejam2barplot() handles a sitenumber parameter

Barplot of ratios of residential population percentages (or other scores) to averages (or other references)
}
\examples{

# Check a long list of indicators for any that are elevated

out <- testoutput_ejamit_100pts_1miles

ejam2barplot(out,
  varnames = names_these_ratio_to_avg,
  main = "Envt & Demog Indicators at Selected Sites Compared to State Averages")
  
ejam2barplot(out,
  varnames = names_these_ratio_to_state_avg,
  main = "Envt & Demog Indicators at Selected Sites Compared to State Averages")

# Residential population percentages only

# vs nationwide avg
ejam2barplot(out)

# vs statewide avg
ejam2barplot(out, 
  varnames = c(names_d_ratio_to_state_avg, names_d_subgroups_ratio_to_state_avg),
  main = "Residential Populations at Selected Sites Compared to State Averages")

# Environmental only

ejam2barplot(out,
  varnames = c(names_e_ratio_to_avg, names_e_ratio_to_state_avg),
  main = "Environmental Indicators at Selected Sites Compared to Averages")
  
  
  ejam2barplot(testoutput_ejamit_100pts_1miles)
  
  plot_barplot_ratios(
    unlist(testoutput_ejamit_1000pts_1miles$results_overall[ , 
    c(..names_d_ratio_to_avg , ..names_d_subgroups_ratio_to_avg) ])
  )

}
\seealso{
\code{\link[=table_xls_format]{table_xls_format()}} \code{\link[=ejam2ratios]{ejam2ratios()}} \code{\link[=ejam2barplot]{ejam2barplot()}}
}
