% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modelDoaggregate.R
\docType{data}
\name{modelDoaggregate}
\alias{modelDoaggregate}
\title{Regression model to predict runtime for doaggregate}
\format{
An object of class \code{lm} of length 12.
}
\usage{
modelDoaggregate
}
\description{
Modeled runtime for doaggregate and ejamit based off 100 runs with random parameters. Use these models to make predictions in app_server
}
\details{
Ejamit's runtime is modeled off radius and number of rows of input dataset, doaggregate runtime is modeled off rows of getblocksnearby output
}
\keyword{datasets}
