% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_speedtest.R
\name{predict_ejamit_runtime}
\alias{predict_ejamit_runtime}
\title{Utility used in app_server to predict ejamit and doaggregate runtimes}
\usage{
predict_ejamit_runtime(rows, radius)
}
\arguments{
\item{rows}{number of locations to be analyzed}

\item{radius}{buffer radius distance, in miles}
}
\description{
Utility used in app_server to predict ejamit and doaggregate runtimes
}
\keyword{internal}
