% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{testdata}
\alias{testdata}
\title{utility to show dir_tree of available files in testdata folders
See list of samples of input files to try in EJAM, and output examples from EJAM functions}
\usage{
testdata(pattern = NULL, installed = TRUE, quiet = FALSE)
}
\arguments{
\item{pattern}{optional query expression, used as regexp passed to \code{\link[fs:dir_ls]{fs::dir_ls()}}}

\item{installed}{If you are a developer who has the local source package,
you can set this parameter to FALSE if you want to work with the
local source package version of the testdata folders
rather than the locally installed version.}

\item{quiet}{set TRUE if you want to just get the path
without seeing all the info in console and without browsing to the folder}
}
\value{
path to local testdata folder comes with the EJAM package
}
\description{
utility to show dir_tree of available files in testdata folders
See list of samples of input files to try in EJAM, and output examples from EJAM functions
}
\examples{
testdata('shapes', quiet = T)
x <- testdata('shape', quiet = TRUE)
x[fs::is_file(x)]

testdata('fips', quiet = T)
testdata('registryid', quiet = T)
testdata("address", quiet = T)

}
\keyword{internal}
