% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testinput_address_table.R
\docType{data}
\name{testinput_address_table}
\alias{testinput_address_table}
\alias{testinput_address_parts1}
\alias{testinput_addresses2}
\alias{testinput_address_table9}
\alias{testinput_addresses_9}
\alias{testinput_address_table_goodnames}
\alias{testinput_address_table_withfull}
\title{datasets for trying address-related functions}
\format{
An object of class \code{data.frame} with 2 rows and 6 columns.
}
\usage{
testinput_address_table
}
\description{
datasets for trying address-related functions (testinput_address_parts1, testinput_addresses2, testinput_address_table9, testinput_addresses_9, testinput_address_table, testinput_address_table_goodnames, testinput_address_table_withfull)
}
\keyword{datasets}
