% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testinput_regid.R
\docType{data}
\name{testinput_regid}
\alias{testinput_regid}
\title{testinput_regid (DATA) test data, vector of EPA FRS Registry ID numbers}
\description{
testinput_regid (DATA) test data, vector of EPA FRS Registry ID numbers
}
\details{
For testing, e.g.,

frs_from_regid(testinput_regid)

mapfast(  frs_from_regid(testinput_regid)  )
}
