% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testinput_xtrac.R
\docType{data}
\name{testinput_xtrac}
\alias{testinput_xtrac}
\title{for internal use}
\format{
An object of class \code{numeric} of length 25.
}
\usage{
testinput_xtrac
}
\description{
for internal use
}
\keyword{datasets}
